requirements.txt
pydns
tqdm
requests